#65 fix: update comment
